Fix OSSMC NamespaceDetailsTab crash and harden namespace Cypress tests (#9952)
* Fix OSSMC NamespaceDetailsTab circular dependency crash

Move getNamespaceRevision helpers out of VirtualList/Renderers into
NamespaceRevisionUtils so NamespaceDetailsOverview no longer triggers
the Config ↔ Renderers TDZ when OSSMC lazy-loads NamespaceDetailsTab.

* Harden Cypress namespace traffic policy tests for OSSMC kiosk mode

Extract shared helpers for namespace detail actions that use the
NamespaceTrafficPolicies confirmation modal, fixing timing issues when
actions are accessed via the minigraph menu in OSSMC.

* Add a potential fix for the occasional Display menu fails.
