+++ frontend/cypress/integration/common/graph_display.ts
@@ -25,7 +25,8 @@ When('user graphs {string} namespaces', (namespaces: string) => {
 });
 
 When('user {string} display menu', (_action: string) => {
-  cy.get('button#display-settings').should('not.be.disabled').click();
+  ensureKialiFinishedLoading();
+  cy.get('button#display-settings').should('be.visible').and('not.be.disabled').click();
 });
 
 When('user enables {string} {string} edge labels', (radio: string, edgeLabel: string) => {
