+++ frontend/cypress/integration/common/sidecar_injection.ts
@@ -1,5 +1,6 @@
 import { Given, Then, When } from '@badeball/cypress-cucumber-preprocessor';
 import { ensureKialiFinishedLoading } from './transition';
+import { confirmNamespaceTrafficPolicyModal, openNamespaceActionsMenu } from './namespace_actions';
 
 // Most of these "Given" implementations are directly using the Kiali API
 // in order to reach a well known state in the environment before performing
@@ -284,54 +285,27 @@ When('I visit the overview page', () => {
   cy.contains('Inbound traffic', { matchCase: false }); // Make sure data finished loading, so avoid broken tests because of a re-render
 });
 
-When('user navigates to the namespace detail page for {string}', function (namespace: string) {
-  this.targetNamespace = namespace;
-  cy.visit({ url: `/console/namespaces/${namespace}?refresh=0` });
-  ensureKialiFinishedLoading();
-});
-
-function openNamespaceActionsMenu(): void {
-  if (Cypress.env('OSSMC')) {
-    cy.intercept('**/api/namespaces/graph*').as('namespaceMinigraph');
-    cy.wait('@namespaceMinigraph');
-    cy.waitForReact();
-    cy.get('button#minigraph-toggle').should('be.visible').click();
-  } else {
-    cy.getBySel('namespace-actions-toggle').should('be.visible').click();
-  }
-}
-
-function confirmAndWaitForNamespacePatch(): void {
-  cy.intercept('PATCH', '**/api/namespaces/**').as('namespacePatch');
-  cy.getBySel('confirm-create').should('be.visible').click();
-  cy.wait('@namespacePatch');
-  ensureKialiFinishedLoading();
-}
-
 When('I override the default automatic sidecar injection policy in the namespace to enabled', function () {
-  ensureKialiFinishedLoading();
   openNamespaceActionsMenu();
   cy.getBySel(`enable-${this.targetNamespace}-namespace-sidecar-injection`).should('be.visible').click();
-  confirmAndWaitForNamespacePatch();
+  confirmNamespaceTrafficPolicyModal();
 });
 
 When(
   'I change the override configuration for automatic sidecar injection policy in the namespace to {string} it',
   function (enabledOrDisabled: string) {
-    ensureKialiFinishedLoading();
     openNamespaceActionsMenu();
     cy.getBySel(`${enabledOrDisabled}-${this.targetNamespace}-namespace-sidecar-injection`)
       .should('be.visible')
       .click();
-    confirmAndWaitForNamespacePatch();
+    confirmNamespaceTrafficPolicyModal();
   }
 );
 
 When('I remove override configuration for sidecar injection in the namespace', function () {
-  ensureKialiFinishedLoading();
   openNamespaceActionsMenu();
   cy.getBySel(`remove-${this.targetNamespace}-namespace-sidecar-injection`).should('be.visible').click();
-  confirmAndWaitForNamespacePatch();
+  confirmNamespaceTrafficPolicyModal();
 });
 
 function switchWorkloadSidecarInjection(enableOrDisable: string): void {
@@ -339,12 +313,9 @@ function switchWorkloadSidecarInjection(enableOrDisable: string): void {
 
   // In OSSMC, the workload actions toggle does not exist. Workload actions are integrated in the minigraph menu
   if (Cypress.env('OSSMC')) {
-    cy.intercept(`**/api/**/workloads/**/graph*`).as('workloadMinigraph');
+    cy.intercept('**/api/**/workloads/**/graph*').as('workloadMinigraph');
     cy.wait('@workloadMinigraph');
-
-    cy.waitForReact();
-
-    cy.get('button#minigraph-toggle').should('be.visible').click();
+    cy.get('button#minigraph-toggle', { timeout: 40000 }).should('be.visible').click();
   } else {
     cy.get('button[data-test="workload-actions-toggle"]').should('be.visible').click();
   }
